testing/web-platform/tests/atomics/window-and-any-worker/wait/good-views/good-views-popup.html

<!doctype html>
<meta charset=utf-8>
<title></title>

<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>

<script>

/*---
description: >
  Test Atomics.wait on arrays that allow atomic operations,
  in an Agent that is allowed to wait.
---*/

function check_done(rs, t) {
  if (rs.length == 7) {
    rs.sort();
    assert_equals(rs[0], "A timed-out");
    assert_equals(rs[1], "B not-equal");
    assert_equals(rs[2], "C not-equal");
    assert_equals(rs[3], "C not-equal");
    assert_equals(rs[4], "C not-equal");
    assert_equals(rs[5], "C not-equal");
    assert_equals(rs[6], "C not-equal");
    t.done();
  }
}

window.addEventListener("load", function() {
  async_test(function(t) {
    let popup = window.open('../empty.html', 'empty');
    let test = this;
    popup.onmessage = function(e) {
      let rs = [];
      let w1 = new Worker('wait-on-zero.js');
      let w2 = new Worker('wait-on-fixed.js');
      let w3 = new Worker('wait-on-index.js');

      w1.onmessage = function(e) {
        rs.push("A " + e.data);
        check_done(rs, t);
      };
      w2.onmessage = function(e) {
        rs.push("B " + e.data);
        check_done(rs, t);
      };
      w3.onmessage = function(e) {
        rs.push("C " + e.data);
        check_done(rs, t);
      };

      let sab = new SharedArrayBuffer(1024);
      let view = new Int32Array(sab, 32, 20);

      w1.postMessage([view]);
      w2.postMessage([view]);

      let good_indices = [
          (view) => 0/-1,
          (view) => '-0',
          (view) => view.length - 1,
          (view) => ({ valueOf: () => 0 }),
          (view) => ({ toString: () => '0', valueOf: false }) // non-callable valueOf triggers invocation of toString
      ];

      setTimeout(function() {
          for (let idxGen of good_indices) {
              let idx = idxGen(view);
              view[idx] = 0;
              // Firefox cannot apply clone algorithm on an Object.
              // See https://developer.mozilla.org/en-US/docs/Web/API/Web_Workers_API/Structured_clone_algorithm
              if (typeof idx == "object") {
                  idx = idx.valueOf ? idx.valueOf() : idx.toString();
              }
              Atomics.store(view, idx, 37);
              w3.postMessage([view, idx]);
          }
      }, 500);
    }
    popup.postMessage([], '*');
  }, "atomics-wait-good-views-from-popup");
});

</script>
